requirements.txt
rich
requests
beautifulsoup4
Pillow
opencv-python
selenium
tensorflow==2.10.0
deepface